+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [unreleased]
+### Fixed
+- Remap chromosome `M` to `MT`, which is the accepted format from ClinVar API
+
 ## [2.8]
 ### Added
 - `validate` endpoint, accepting germline or somatic json submissions and supporting validation against the official schema